Update dcraw_common.cpp
The adobe_coeff function has a bug where the color values are written to out of bounds of the cam_xyz array. The array is declared as cam_xyz[4][3] whereas the inner for loop is looping through indexes 0 to 11 (the j variable). This variable is then used to set the value in cam_xyz[0][j] where j > 2 meaning out of bounds.